(game) rename board.js to match.js
'match' seems to be a better name, because a 'game' has more things
than just a current match. For example, we can add statistics to the
game, named players, custom things to pass to a game UI renderer,
and other stuff. 'match' is just the tic-tac-toe rules, played by
two generic players.